fix: add overflow and zero-size checks for UEFI memory descriptors
Add validation for zero-sized memory descriptors and use safe arithmetic
to prevent overflow when calculating descriptor addresses. Also mark
framebuffer operations as unsafe where appropriate.

+++ nitrogen/src/iwlwifi/connection_state.rs
@@ -280,7 +280,9 @@ pub fn try_init_wifi_device_step() {
             let mac_acquired = {
                 let ctx = WIFI_INIT_CTX.lock();
                 let health = ctx.health.as_ref();
-                match unsafe { mmio::checked_read_u32(mmio.add(CSR_GP_CNTRL as usize), health) } {
+                match unsafe {
+                    mmio::checked_read_u32(mmio.add(CSR_GP_CNTRL as usize) as usize, health)
+                } {
                     mmio::SafeReadResult::Value(v) if v & CSR_GP_CNTRL_MAC_CLOCK_READY != 0 => true,
                     mmio::SafeReadResult::MasterAbort | mmio::SafeReadResult::DeviceGone => {
                         mmio::disarm_mmio_watchdog();
@@ -315,7 +317,10 @@ pub fn try_init_wifi_device_step() {
                                     mmio::arm_mmio_watchdog(0, pci_bdf, bridge_bdf);
                                 }
                                 let clock_ready = match unsafe {
-                                    mmio::checked_read_u32(mmio.add(CSR_GP_CNTRL as usize), health)
+                                    mmio::checked_read_u32(
+                                        mmio.add(CSR_GP_CNTRL as usize) as usize,
+                                        health,
+                                    )
                                 } {
                                     mmio::SafeReadResult::Value(v)
                                         if v & CSR_GP_CNTRL_MAC_CLOCK_READY != 0 =>
